refactor: extract Profile; delete fresh.ts
